@@ -41,101 +44,105 @@
     }
   },
   "dependencies": {
-    "@babel/cli": "^7.4.3",
-    "@babel/core": "^7.5.5",
-    "@babel/plugin-proposal-class-properties": "^7.4.0",
-    "@babel/plugin-proposal-decorators": "^7.4.0",
+    "@babel/cli": "^7.6.4",
+    "@babel/core": "^7.6.4",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/plugin-proposal-decorators": "^7.6.0",
     "@babel/plugin-proposal-export-default-from": "^7.5.2",
-    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
-    "@babel/plugin-proposal-function-sent": "^7.0.0",
-    "@babel/plugin-proposal-json-strings": "^7.0.0",
-    "@babel/plugin-proposal-numeric-separator": "^7.0.0",
-    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-    "@babel/plugin-syntax-import-meta": "^7.0.0",
-    "@babel/plugin-transform-runtime": "^7.5.5",
-    "@babel/preset-env": "^7.4.5",
+    "@babel/plugin-proposal-export-namespace-from": "^7.5.2",
+    "@babel/plugin-proposal-function-sent": "^7.5.0",
+    "@babel/plugin-proposal-json-strings": "^7.2.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.2.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-syntax-import-meta": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.6.2",
+    "@babel/preset-env": "^7.6.3",
     "@babel/preset-flow": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-react": "^7.6.3",
     "@babel/preset-stage-3": "^7.0.0",
-    "@babel/preset-typescript": "^7.1.0",
-    "@babel/register": "^7.5.5",
-    "@babel/runtime": "^7.5.5",
+    "@babel/preset-typescript": "^7.6.0",
+    "@babel/register": "^7.6.2",
+    "@babel/runtime": "^7.6.3",
     "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "10.0.2",
-    "babel-jest": "^24.8.0",
-    "babel-loader": "^8.0.0-beta.6",
+    "babel-eslint": "10.0.3",
+    "babel-jest": "^24.9.0",
+    "babel-loader": "^8.0.6",
     "babel-plugin-css-modules-transform": "^1.6.2",
-    "babel-plugin-emotion": "^10.0.19",
+    "babel-plugin-emotion": "^10.0.21",
     "babel-polyfill": "^6.26.0",
     "babel-preset-react-boilerplate": "1.1.1",
     "babel-preset-react-hmre": "1.1.1",
-    "body-parser": "^1.18.3",
+    "body-parser": "^1.19.0",
     "classnames": "^2.2.6",
     "connected-react-router": "^6.5.2",
     "cookie-parser": "^1.4.4",
-    "core-js": "^3.0.1",
+    "core-js": "^3.3.3",
     "cors": "^2.8.5",
-    "cross-env": "^5.2.0",
-    "d3": "^5.7.0",
-    "dotenv": "^8.0.0",
-    "es6-promise": "^4.2.6",
-    "express": "^4.16.4",
-    "git-revision-webpack-plugin": "^3.0.3",
-    "history": "^4.7.2",
+    "cross-env": "^6.0.3",
+    "d3": "^5.12.0",
+    "dotenv": "^8.2.0",
+    "es6-promise": "^4.2.8",
+    "express": "^4.17.1",
+    "git-revision-webpack-plugin": "^3.0.4",
+    "history": "^4.10.1",
     "isomorphic-fetch": "^2.2.1",
-    "next": "^7.0.2",
+    "next": "^9.1.1",
     "openssl": "^1.1.0",
-    "prop-types": "^15.6.2",
+    "prop-types": "^15.7.2",
     "raf": "^3.4.1",
-    "raw-loader": "^3.0.0",
-    "react": "^16.8.3",
-    "react-app-polyfill": "^1.0.1",
-    "react-dom": "^16.8.3",
-    "react-helmet": "^5.2.0",
-    "react-hot-loader": "^4.3.12",
+    "raw-loader": "^3.1.0",
+    "react": "^16.11.0",
+    "react-app-polyfill": "^1.0.4",
+    "react-dom": "^16.11.0",
+    "react-helmet": "^5.2.1",
+    "react-hot-loader": "^4.12.15",
     "react-redux": "^7.1.1",
-    "react-router-dom": "^4.3.1",
-    "react-router-redux": "^5.0.0-alpha.9",
-    "redux": "^4.0.1",
-    "redux-devtools-extension": "^2.13.5",
+    "react-router-dom": "^5.1.2",
+    "redux": "^4.0.4",
+    "redux-devtools-extension": "^2.13.8",
     "redux-thunk": "^2.3.0",
     "request": "^2.88.0",
     "sockjs": "^0.3.19",
-    "sockjs-client": "^1.3.0",
+    "sockjs-client": "^1.4.0",
     "string-replace-webpack-plugin": "^0.1.3",
     "uuid": "^3.3.3"
   },
   "devDependencies": {
-    "copy-webpack-plugin": "^4.6.0",
-    "css-loader": "1.0.1",
-    "eslint": "^6.3.0",
-    "eslint-config-airbnb": "17.1.1",
-    "eslint-loader": "2.2.1",
+    "copy-webpack-plugin": "^5.0.4",
+    "css-loader": "3.2.0",
+    "eslint": "^6.5.1",
+    "eslint-config-airbnb": "18.0.1",
+    "eslint-loader": "3.0.2",
     "eslint-plugin-import": "2.18.2",
     "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-react": "7.14.3",
-    "eslint-plugin-react-hooks": "^2.1.0",
+    "eslint-plugin-react": "7.16.0",
+    "eslint-plugin-react-hooks": "^2.2.0",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "file-loader": "2.0.0",
+    "file-loader": "4.2.0",
     "html-webpack-plugin": "3.2.0",
-    "husky": "1.1.4",
-    "jest": "^24.8.0",
-    "jest-cli": "^24.8.0",
-    "node-hot-loader": "^1.12.3",
-    "node-sass": "^4.10.0",
-    "nodemon": "^1.18.9",
+    "husky": "3.0.9",
+    "jest": "^24.9.0",
+    "jest-cli": "^24.9.0",
+    "node-hot-loader": "^1.17.6",
+    "node-sass": "^4.12.0",
+    "nodemon": "^1.19.4",
     "react-addons-test-utils": "^15.6.2",
-    "react-test-renderer": "16.6.3",
-    "regenerator-runtime": "^0.13.2",
-    "rimraf": "2.6.3",
-    "sass-loader": "7.1.0",
-    "style-loader": "0.23.1",
-    "url-loader": "1.1.2",
-    "webpack": "^4.29.6",
-    "webpack-cli": "^3.3.0",
-    "webpack-dev-server": "^3.2.1",
+    "react-test-renderer": "16.11.0",
+    "regenerator-runtime": "^0.13.3",
+    "rimraf": "3.0.0",
+    "sass-loader": "8.0.0",
+    "style-loader": "1.0.0",
+    "url-loader": "2.2.0",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.9",
+    "webpack-dev-server": "^3.9.0",
     "webpack-shell-plugin": "^0.5.0"
+  },
+  "nodemonConfig": {
+    "ignore": [
+      "*.hot-update.json"
+    ]
   },
   "husky": {
     "hooks": {
