@@ -1,29 +1,29 @@
-import React from 'react';
-
+import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux'
 import Head from './head'
 
-class Home extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {}
-  }
+const Home = () => {
+  const [counter, setCounterNew] = useState(0)
 
-  render() {
-    return (
-      <div>
-        <Head title="Hello" />
-        <div> Hello World Home </div>
-      </div>
-    )
-  }
+  return (
+    <div>
+      <Head title="Hello" />
+      <button
+        type="button"
+        onClick={() => setCounterNew(counter + 1)}
+      >
+        updateCounter
+      </button>
+      <div> Hello World Dashboard {counter} </div>
+    </div>
+  )
 }
 
 Home.propTypes = {}
 
 const mapStateToProps = () => ({})
 
-const mapDispatchToProps = dispatch => bindActionCreators({}, dispatch)
+const mapDispatchToProps = (dispatch) => bindActionCreators({}, dispatch)
 
 export default connect(mapStateToProps, mapDispatchToProps)(Home)
