@@ -1,3 +1,4 @@
+/* eslint-disable camelcase */
 import React from 'react'
 import PropTypes from 'prop-types'
 import { bindActionCreators } from 'redux'
@@ -7,7 +8,7 @@ import { withRouter } from 'react-router-dom';
 // const cookies = new Cookies();
 
 class Startup extends React.Component {
-  componentWillMount() { }
+  UNSAFE_componentWillMount() { }
 
   render() {
     return this.props.children;
@@ -27,6 +28,6 @@ Startup.defaultProps = {
 const mapStateToProps = () => ({
 })
 
-const mapDispatchToProps = dispatch => bindActionCreators({ }, dispatch)
+const mapDispatchToProps = (dispatch) => bindActionCreators({ }, dispatch)
 
 export default withRouter(connect(mapStateToProps, mapDispatchToProps)(Startup))
