@@ -1,31 +1,23 @@
-import React from 'react';
+import React, { useState } from 'react';
 
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux'
 import Head from './head'
 
-class Dummy extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      counter: 0
-    }
-  }
-
-  render() {
-    return (
-      <div>
-        <Head title="Hello" />
-        <div> Hello World {this.state.counter} </div>
-      </div>
-    )
-  }
+const Dummy = () => {
+  const [counter] = useState(0)
+  return (
+    <div>
+      <Head title="Hello" />
+      <div> Hello World {counter} </div>
+    </div>
+  )
 }
 
 Dummy.propTypes = {}
 
 const mapStateToProps = () => ({})
 
-const mapDispatchToProps = dispatch => bindActionCreators({}, dispatch)
+const mapDispatchToProps = (dispatch) => bindActionCreators({}, dispatch)
 
 export default connect(mapStateToProps, mapDispatchToProps)(Dummy)
